+++ bot/src/__init__.py
@@ -1,0 +1,4 @@
+from .config import TELEGRAM_API_ID, TELEGRAM_API_HASH, CENSOR, STRING, censor_list
+from .telegram_client import get_telegram_client
+from .history import get_chat_history
+from .message_utils import filter_bot_messages, count_words, censor_result
