+++ bot/src/config.py
@@ -1,0 +1,9 @@
+import os
+
+TELEGRAM_API_ID = os.environ.get("TELEGRAM_API_ID")
+TELEGRAM_API_HASH = os.environ.get("TELEGRAM_API_HASH")
+CENSOR = os.environ.get("CENSOR")
+STRING = os.environ.get("STRING")
+
+# List of words to censor. Generated by splitting the CENSOR environment variable by commas
+censor_list = CENSOR.split(",")
